+++ src/features/staff-walkthrough-improvements/edit-walkthrough/tinymce/fix-floating-menus.ts
@@ -1,3 +1,4 @@
+import { subscribe } from '@ta-x-components';
 import { Constants } from '@ta-x-globals';
 import { getElementCoordinates, waitForElement } from '@ta-x-utilities';
 
@@ -11,6 +12,7 @@ const setTopPosition = (): void => {
 
 const listen = (): void => {
   window.addEventListener('scroll', setTopPosition);
+  subscribe('tinymce:repositionFloatingMenus', setTopPosition);
 };
 
 export const fixFloatingMenus = async(container: HTMLElement): Promise<void> => {